Progress with OpenGL interop! First working version.
Importing the Vulkan buffer into OpenGL works, and i
can render a simple background color and a nice little
rotating cat picture.

OpenGL renders the cat picture to the imported VK buffer,
Vulkan textures from it to display the OpenGL rendered
image on the sides of the rotating 3D cube.

As synchronization via semaphores is not implemented yet,
all this happens without explicit handshaking between GL and
Vulkan. Only a glFinish() is used at end of OpenGL rendering.
Also proper Vulkan->OpenGL->Vulkan transitioning is missing.
So this may not work on all driver+gpu combos yet, and also
not with maximum performance.

However: We do get properly rendered OpenGL animations
on the sides of the displayed rotating Vulkan cube, and
with smooth high framerate on a 85Hz monitor, and without
any malfunctions, artifacts, or error messages in dmesg.

-> First working demonstration of OpenGL->Vulkan interop on
   a fullscreen direct mode display (RandR leased output) on
   Mesa 19.2.1 gallium radeonsi + amdvlk open-source driver,
   vulkan api version 1.1.127, driver version 2019Q4-3 on
   Linux 5.4.0-rc7+ and Linux 5.3.0-26 distro kernel on
   Ubuntu 19.10 with XOrg 1.20.5. As this is also what the
   latest HWE stack is shipping, it should also work on
   Ubuntu 18.04.4-LTS HWE.

   Hardware is AMD Ryzen (Vega 11, DCN-1) on a single-display,
   single X-Screen setup under KDE 5 Plasma.

   Currently we only have RGBA8 framebuffers available under
   Vulkan display direct mode, but then this is not a HDR
   display, but a 20 year old VGA CRT. Maybe it makes a
   difference?