Windows: Attach VkSurfaceFullScreenExclusiveWin32InfoEXT.
This is mandated by the spec for fullscreen exclusive mode for
surfaces the way we create them. The AMD drivers seem to be
lenient and didn't complain about our lack of this, but the
Windows-10 NVidia driver failed swapchain creation, presumably
because we were missing this. Repeat of test tbd.