Windows: Partial fixup for NVidia driver.
The NVidia driver on Windows 10 doesn't like it if the
window used for attaching the Vulkan swapchain also
has an OpenGL rendering context attached. AMD driver
did not care.

Create a 2nd "dummy window" for the sole purpose of
attaching the OpenGL context to it, then make that
context current to the thread which will execute both
OpenGL and Vulkan code. Now Vulkan swapchain creation
works, because the fullscreen window for Vulkan output
does not have a OpenGL context attached.

Regular Vulkan rendering without interop works.
OpenGl>Vulkan interop shows roughly correct results,
but is a bit broken, with pixel trash in the rendered cube.
I assume / hope this is because we don't use proper
semaphores yet to sync OpenGL and Vulkan rendering for
proper handshaking, and because we don't use actual
Vulkan command buffers for transitioning between OpenGL
and Vulkan rendering. While the AMD driver on Windows and
Linux didn't care and things just worked, this seems to be not
the case for NVidia at least on Windows-10.

-> More work to do. Baby steps...