Set color values for cube to something more reasonable for PQ
If PQ is signalled by the driver to the display, due to our HDR10_ST2084
colorspace selection, then we should also set something reasonable for
the PQ range.

I'm undecided if this actually works? Results on Windows are mixed, but
not clear if i'm mistaken, the AMD driver is, or the monitor is acting up?
Rechecking on Linux should help, at least we know what is sent to the
monitor there - or so i hope.