Random new set of color values for testing.
Now the OpenGL colors are expressed in BT 2020, unit is
nits. Testing suggests that at least with local dimming
disabled, the monitor does map the given nits range 0 -
600 nits to something like maximally dark to maximum
brightness. This would be at least consistent with the
monitor reporting 0.049 - 603 nits range.

Question is if all the non-sensical stuff is see is due
to the monitor misbehaving (it seems to be quite buggy
in many ways), or if what i see is just the state of the
art, which would be quite sad. Or some misunderstanding
on my side.