Factor out HDR meta data setting, adapt to Vesa CTS defaults.
Put the setting code into a separate function, and for the moment,
code a color gamut == display reported native gamut, so the display
should not need to gamut remap.

Try to select luminance levels for min/max/maxCLL/maxFALL, so the
display does not have to tone map either.

Select cube luminance as the maxFALL of my Samsung monitor, so
it has a chance to actually reproduce the light levels specified.

-> Monitor is still not reproducing the specified light levels,
   shows substantially lower levels.