Switch from rotating cube to 1:1 draw of screen-aligned quad.
Enough with the cubes! Make the minimal change to render a single
screen-aligned quad instead of a cube, and a OpenGL interop texture
applied to that quad, of the size of the screen, so we can blit
OpenGL rendered content directly to the swapchain, only applying
the PQ OETF during the blit.

This is the most basic PoC for the kind of rendering we actually
will need within PTB. It can be simplified, and maybe also replaced
by a direct blit from the OpenGL texture backing VKImage into the
swapchain fb, or ideally even by exporting the swapchain framebuffer
images themselves as backing memory for the OpenGL textures, so
OpenGL could directly render to the swapchain framebuffer - a zero
copy essentially. Have to learn more about Vulkan to see if/how this
can be done.

-> Tested on AMD + Linux.