Add optimized path for interop-texture -> swapchain image copy.
By default (if --useshader cmd line option is omitted), we now
use vkCmdCopyImage() for an efficient (memcpy() style) copy from
the VkImage representing the OpenGL interop texture to the swapchain
image representing the Vulkan backbuffer. For that we need to do
some layout transitions on the source image and swapchain image.

The 3D rendering engine is not used in this case.

Option --useshader will use the old pass that uses the 3D engine
to draw a screen aligned, textured, fullscreen quad, where a
Vulkan fragment shader reads the interop texture and writes it
one-to-one texel->pixel mapped into the swapchain image. This
requires way more setup code than the simple copy, but would
allow pixel format conversions or shader application.

Note that atm. the Vulkan shader is just a simple passthrough
shader. All HDR post-processing is done on the OpenGL side with
a GLSL fragment shader. This matches the way things will operate
in Psychtoolbox, where we have a convenient and flexible OpenGL
GLSL based image post-processing pipeline.

Ideally we would do without an interop texture VKImage and attach
the Vulkan swapchain images directly as OpenGL interop textures,
saving this extra identity copy/renderpass on the Vulkan side.
Unfortunately i couldn't find any way to get to the VKDeviceMemory
object which backs swapchain images, and therefore i can't use the
interop extensions to export the swapchain image backing memory as
a texture to OpenGL for direct rendering into it. Could not find a
hint on anything usable in the Spec, or by grep'ing through the AMD
driver, so i guess this exotic use case is not supported atm.

The expected extra time needed for this copy pass is about 0.25 msecs
to a bit less than 2 msecs, depending on gpu type (IGP or discrete)
and memory speed. However, the estimate for a 500 Euros PC with an
AMD Ryzen 5 2400G Raven Ridge APU and about 40 GB/sec memory bandwidth
suggests below 1 msec of time penalty for a 2560x1440 HDR monitor, so
i think this should be acceptable. Under AMD RADV, the demo can sustain
144 fps without frame drops in this simplified synthetic scenario.