Optimize transfer from OpenGL interop image to swapchain image.
If the pixel format of source and destination image match, use
vkCmdCopyImage() like before, to essentially implement a simple
and efficient GPU memcpy() of raw data.

On mismatched formats which require a Red<=>Blue channel swizzle,
use vkCmdBlitImage() instead, which can do this, but may be a bit
less efficient, depending on how the swizzle is implemented in hw.

=> Now we always get an as efficient as possible copy pass from
the OpenGL interop texture source VKImage to the  destination
swapchain VKImage.

=> Note that in case of AMD, where the OpenGL implementation can
only render into linear textures, but the swapchain is supposed
to by in tiled format, the copy will have to do a tiling during
copy, which may cost a bit of performance. How big the impact is
is unknown so far. Ideally AMD's OpenGL implementation would be
extended to support optimal tiling, ie. a tiled texture for rendering.